bug 1005013 - move newrelic conf to nconf
