+++ lib/kumascript/index.js
@@ -6,7 +6,7 @@
 /*jshint node: true, expr: false, boss: true */
 
 var _ = require('underscore');
-    submodules = ['server', 'caching', 'loaders', 'templates', 'macros', 'api',
+    submodules = ['server', 'caching', 'conf', 'loaders', 'templates', 'macros', 'api',
                   'errors', 'repl', 'utils', 'test-utils'];
 
 for (var i=0,n; n=submodules[i]; i++) {
