+++ lib/kumascript/server.js
@@ -4,7 +4,10 @@
 
 /*jshint node: true, expr: false, boss: true */
 
-if (process.env.NEW_RELIC_LICENSE_KEY) require('newrelic');
+var ks_conf = require(__dirname + '/conf');
+
+var newrelic_conf = ks_conf.nconf.get('newrelic');
+if (newrelic_conf && newrelic_conf.license_key) require('newrelic');
 
 // ### Prerequisites
 var util = require('util'),
