+++ newrelic.js
@@ -4,20 +4,10 @@
  * See lib/config.defaults.js in the agent distribution for a more complete
  * description of configuration variables and their potential values.
  */
-exports.config = {
-  /**
-   * Note: We rely on the following config environment variables:
-   * NEW_RELIC_APP_NAME
-   * NEW_RELIC_LICENSE_KEY
-   * NEW_RELIC_HIGH_SECURITY
-   */
 
-  logging: {
-    /**
-     * Level at which to log. 'trace' is most useful to New Relic when diagnosing
-     * issues with the agent, 'info' and higher will impose the least overhead on
-     * production applications.
-     */
-    level: 'info'
-  }
-}
+// Load from nconf to keep conf values in a single place
+var ks_conf = require(__dirname + '/lib/kumascript/conf');
+
+var newrelic_conf = ks_conf.nconf.get('newrelic');
+
+exports.config = newrelic_conf;
